feat(smart-router): allow a poll cadence slower than the chain's block time (MAG-2985)
--chain-tracker-poll-divisor was capped at 1, so the slowest reachable cadence
was one poll per block time. That floor was justified by an argument that does
not hold: chainstate.StalenessWindow (max(10 x avgBlockTime, 2s)) does not move
with this knob. The knob moves how long a tip can go unrefreshed, which is the
OTHER side of that comparison — and at one poll per block time it was sitting an
order of magnitude inside a window it was said to be approaching.

Widen the range to [0.25, 8]. At 0.25 the tracker polls once per four block
times: 8x less than the built-in cadence and 4x less than the old floor. Measured
per chain, requests per minute per endpoint:

  APT1      600 -> 75     SOLANA    300 -> 37.5
  BASE       60 -> 7.5    ETH1      9.2 -> 1.2

The value is fractional, so the knob is a float end to end: the cobra flag
(Float64), the viper read (GetFloat64), the config field, and the division. The
division is the one that had to change shape rather than type —
time.Duration(divisor) truncates 0.25 to 0 and panics — so it now divides in
float64 and converts after. GetInt was the quieter hazard: it would have
truncated 0.25 to 0, which is the "unset" sentinel, so an operator asking for the
largest relief would have silently received the default instead.

NaN and +Inf are rejected explicitly. NaN fails every ordinary comparison, so it
would pass a range check and make the interval NaN, which time.Duration renders
as a large negative — a timer that fires immediately and hot-loops the upstream.

What the fractional low end does NOT change is the staleness window, and the two
common cases stay comfortably inside it: an idle endpoint is refreshed only by
its own poll, so its gap IS the interval (4 x avgBlockTime at 0.25, against a 10x
window), and a served endpoint is refreshed by relay harvest regardless of
cadence. The exposure is the seam between them — an endpoint that trips the
traffic gate and then goes quiet is refreshed by neither, and the worst-case gap
becomes (skips+1) x interval, which at 0.25 is 20 x avgBlockTime.

That hazard is a property of the PRODUCT of this knob and the gate's skip budget,
not of either alone, which is exactly why a range check on either cannot catch
it. warnIfCadenceOutrunsStaleness reports it once per chain at startup rather
than this range pretending to prevent it, and it warns rather than rejects: a
configuration that is safe for both common cases should not be refused for a
seam. chaintracker.DefaultMaxRelaySkipsBeforePoll is exported so the check reads
the real budget instead of duplicating the number.

Verified against a real node (LAVA, 15s blocks), one endpoint, no relay traffic:

  default            PollIntervalMs=7500    6 polls in 60s   no warning
  divisor 0.25       PollIntervalMs=60000   1 poll  in 60s   warning fired

+++ protocol/rpcsmartrouter/rpcsmartrouter_server.go
@@ -230,7 +230,10 @@ func (rpcss *RPCSmartRouterServer) ServeRPCRequests(
 		// Dedicated-poll cadence, read from viper for the same reason as the flag above (an
 		// embedded server with no flags bound gets 0, which IS the default). endpointstate
 		// validates the value and reverts an out-of-range one.
-		PollIntervalDivisor: viper.GetInt(endpointstate.PollDivisorFlagName),
+		// GetFloat64, not GetInt: the divisor is fractional at the low end, and GetInt would
+		// truncate 0.25 to 0 — which is the "unset" sentinel, so an operator asking for the
+		// largest relief would silently receive the default instead.
+		PollIntervalDivisor: viper.GetFloat64(endpointstate.PollDivisorFlagName),
 		// Feed every positive poll/relay block into the per-chain tip (cheap monotonic write)
 		// and mirror the resulting guarded tip into the router-wide latest-block gauge (MAG-2629).
 		OnTipObservation: rpcss.onTipObservation,
